(BSR)[API] fix: Do not require `pytest` outside of local environment and tests
As per the previous commit, we require `factory-boy` everywhere.
However, we can do without `pytest` by splitting the `core.testing`
module and moving factory-related code to `core.factories`.
