+++ api/src/pcapi/core/criteria/factories.py
@@ -1,6 +1,6 @@
 import factory
 
-from pcapi.core.testing import BaseFactory
+from pcapi.core.factories import BaseFactory
 
 from . import models
