+++ api/src/pcapi/core/providers/factories.py
@@ -6,10 +6,10 @@ import secrets
 
 import factory
 
+from pcapi.core.factories import BaseFactory
 from pcapi.core.offerers import factories as offerers_factories
 from pcapi.core.offerers import models as offerers_models
 import pcapi.core.providers.repository as providers_repository
-from pcapi.core.testing import BaseFactory
 from pcapi.utils.crypto import encrypt
 
 from . import models
